requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev.txt requirements/dev.in
#
astroid==1.4.8            # via pylint
automaton==1.4.0
behave==1.2.5
click==6.6                # via pip-tools
debtcollector==1.8.0      # via automaton
first==2.0.1              # via pip-tools
flake8==3.0.4
isort==4.2.5              # via pylint
lazy-object-proxy==1.2.2  # via astroid
lxml==3.6.4
mccabe==0.5.2             # via flake8, pylint
paho-mqtt==1.2
parse-type==0.3.4         # via behave
parse==1.6.6              # via behave, parse-type
pbr==1.10.0               # via automaton, debtcollector
pip-tools==1.7.0
pluggy==0.3.1             # via tox
PrettyTable==0.7.2        # via automaton
py==1.4.31                # via pytest, tox
pycodestyle==2.0.0        # via flake8
pyflakes==1.2.3           # via flake8
pylint==1.6.4
pytest==3.0.1
PyYAML==3.11
six==1.10.0               # via astroid, automaton, behave, debtcollector, parse-type, pip-tools, pylint
tox==2.3.1
virtualenv==15.0.3        # via tox
wrapt==1.10.8             # via astroid, debtcollector
yapf==0.11.1
